feat: Implemented session-based access hash(HMACSHA256) for Users and Channels
1. Access hashes are no longer stored in the database but are generated
dynamically.
2. Access hashes are only valid within the current session
(bound to the permanent auth key).

+++ source/src/MyTelegram.Services/Services/QueuedObjectMessageSender.cs
@@ -37,7 +37,7 @@ public class QueuedObjectMessageSender(
         List<long>? excludeUserIds = null
         ) where TData : IObject
     {
-        sessionMessageQueueProcessor.Enqueue(new LayeredPushMessageCreatedIntegrationEvent((int)peer.PeerType,
+        sessionMessageQueueProcessor.Enqueue(new LayeredPushMessageCreatedIntegrationEvent(peer.PeerType,
                 peer.PeerId,
                 data.ToBytes(),
                 excludeAuthKeyId,
@@ -89,7 +89,7 @@ public class QueuedObjectMessageSender(
         }
 
         sessionMessageQueueProcessor.Enqueue(new LayeredPushMessageCreatedIntegrationEvent<TExtraData>(
-                (int)peer.PeerType,
+               peer.PeerType,
                 peer.PeerId,
                 data.ToBytes(),
                 excludeAuthKeyId,
